refactor: separate fns to service
